Update package dependencies and refactor MarkdownContent component to improve code readability and usability. Replace toast notification library with 'sonner', implement copy-to-clipboard functionality, and enhance chat state management with updated types and client creation method.
