+++ package.json
@@ -30,11 +30,12 @@
     "@supabase/supabase-js": "^2.49.8",
     "@tremor/react": "^3.18.7",
     "@vapi-ai/web": "^2.3.1",
+    "@vercel/ai": "^1.0.0",
     "ai": "^4.3.16",
     "class-variance-authority": "^0.7.1",
     "cloudinary": "^2.6.1",
     "clsx": "^2.1.1",
-    "date-fns": "^4.1.0",
+    "date-fns": "^3.6.0",
     "elevenlabs": "^1.59.0",
     "elevenlabs-node": "^2.0.3",
     "framer-motion": "^12.15.0",
@@ -55,17 +56,16 @@
     "zod": "^3.25.34"
   },
   "devDependencies": {
-    "@eslint/eslintrc": "^3",
-    "@tailwindcss/postcss": "^4",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "@types/react-syntax-highlighter": "^15.5.13",
-    "eslint": "^9",
+    "eslint": "^8.56.0",
     "eslint-config-next": "15.3.2",
     "supabase": "^2.23.4",
-    "tailwindcss": "^4",
-    "tw-animate-css": "^1.3.0",
+    "tailwindcss": "^3.4.1",
+    "postcss": "^8.4.35",
+    "autoprefixer": "^10.4.17",
     "typescript": "^5"
   }
 }
