+++ src/lib/supabaseClient.ts
@@ -1,6 +1,6 @@
 // src/lib/supabaseClient.ts
 // This file is EXCLUSIVELY for CLIENT-SIDE Supabase access.
-import { createPagesBrowserClient } from '@supabase/auth-helpers-nextjs';
+import { createClientComponentClient } from '@supabase/auth-helpers-nextjs';
 
 const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
 const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY;
@@ -9,4 +9,4 @@ if (!supabaseUrl || !supabaseAnonKey) {
   throw new Error("Supabase URL or Anon Key is missing from .env.local. Make sure NEXT_PUBLIC_SUPABASE_URL and NEXT_PUBLIC_SUPABASE_ANON_KEY are set.");
 }
 
-export const supabase = createPagesBrowserClient(supabaseUrl, supabaseAnonKey);
+export const supabase = createClientComponentClient();
