+++ src/types/chat.ts
@@ -1,10 +1,18 @@
-import { VercelAIMessage } from '@vercel/ai';
+import { Message as VercelAIMessage } from 'ai';
 
-export interface ChatMessage extends VercelAIMessage {
+export interface ChatMessage extends Omit<VercelAIMessage, 'createdAt'> {
   conversationId?: string;
-  createdAt?: string;
+  createdAt?: Date;
 }
 
 export interface ChatMessageWithId extends ChatMessage {
   id: string;
+}
+
+export interface Conversation {
+  id: string;
+  title: string;
+  messages: ChatMessageWithId[];
+  createdAt: string;
+  updatedAt: string;
 }
